Convert Compendium data into a DataFrame
